requirements.txt
click==7.1.2
dominate==2.6.0
Flask==1.1.2
Flask-Bootstrap==3.3.7.1
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
greenlet==1.0.0
importlib-metadata==4.0.1
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
python-dateutil==2.8.1
six==1.15.0
SQLAlchemy==1.3.19
typing-extensions==3.7.4.3
visitor==0.1.3
Werkzeug==1.0.1
WTForms==2.3.3
zipp==3.4.1
